create RM with token
