+++ components/modals/UpdateStepFormModal.tsx
@@ -1,12 +1,13 @@
 import { Dialog } from '@headlessui/react'
 import { SetStateAction } from 'react'
-import { StepWithLib } from '~/pages/roadmaps/create'
 import { LibraryInfo } from '~/server/types'
 import { Library } from '$prisma/client'
 import { SelectOption } from '../parts/SelectInput'
 import UpdateStepForm from '../roadmaps/step/UpdateStepForm'
+import { StepWithLib } from '../roadmaps/Roadmap'
 
 type Props = {
+  token: string
   isOpen: boolean
   setIsOpen: React.Dispatch<SetStateAction<boolean>>
   libTitleOptions: SelectOption[]
@@ -14,6 +15,7 @@ type Props = {
   handleLibInputChange: (keyword: string) => void
   onSearchLibraries: (keyword: string) => Promise<Library[]>
   onCreateLibrary: (
+    token: string,
     title: string,
     link?: string | null | undefined
   ) => Promise<Library>
@@ -23,6 +25,7 @@ type Props = {
 }
 
 const UpdateStepFormModal = ({
+  token,
   isOpen,
   setIsOpen,
   handleLibInputChange,
@@ -49,6 +52,7 @@ const UpdateStepFormModal = ({
         ステップを編集
       </Dialog.Title>
       <UpdateStepForm
+        token={token}
         handleLibInputChange={handleLibInputChange}
         onSearchLibraries={onSearchLibraries}
         onCreateLibrary={onCreateLibrary}
